UEFI 64-bit also works

+++ basic-uefi-image.sh
@@ -1,3 +1,10 @@
+#!/bin/bash
+
+if [ $# -ne 1 ]; then
+	echo Usage: $0 BOOT[IA32/x64].EFI
+	exit 1
+fi
+
 # UEFI FAT partition with bootable EFI application.
 dd if=/dev/zero of=uefi.img bs=512 count=262144
 parted uefi.img -s "mktable gpt"
@@ -9,7 +16,7 @@ mkfs.vfat -n UEFI-TOOLS fat.img
 
 mmd -i fat.img ::/EFI
 mmd -i fat.img ::/EFI/BOOT
-mcopy -i fat.img BOOTIA32.EFI ::/EFI/BOOT/BOOTIA32.EFI
+mcopy -i fat.img $1 ::/EFI/BOOT/$1
 
 dd if=fat.img of=uefi.img bs=512 seek=2048 count=131072 conv=notrunc
